Struct btreelist::BTreeList

pub struct BTreeList<T, const B: usize = 6> { /* private fields */ }

A list with efficient insert and removal in the middle.

It may be worth benchmarking your use case and trying to use a Box<T> instead of a plain T as this can improve performance in some cases. Similar word-length wrapper types would also work e.g. Rc.

let mut list = BTreeList::default();
list.push(1);
list.push(2);

assert_eq!(list.len(), 2);
assert_eq!(list[0], 1);

assert_eq!(list.pop(), Some(2));
assert_eq!(list.len(), 1);

list[0] = 7;
assert_eq!(list[0], 7);

list.extend([1,2,3]);

for x in &list {
    println!("{x}");
}
assert_eq!(list, btreelist![7, 1, 2, 3]);

Implementations

impl<T, const B: usize> BTreeList<T, B>

pub fn new() -> Self

Construct a new, empty BTreeList.

No allocation occurs until elements are added.

// create a BTreeList with the default B parameter
let mut list : BTreeList<i32> = BTreeList::new();
// create a BTreeList with a custom B parameter
let mut list : BTreeList<i32, 32> = BTreeList::new();

pub fn len(&self) -> usize

Get the length of the list.

let list = btreelist![1, 2, 3];
assert_eq!(list.len(), 3);

pub fn is_empty(&self) -> bool

Check if the list is empty.

let mut list: BTreeList<_> = BTreeList::new();
assert!(list.is_empty());

list.push(1);
assert!(!list.is_empty());

pub fn iter(&self) -> Iter<'_, T, B>

Create an iterator through the list.

let x = btreelist![1, 2, 4];
let mut iterator = x.iter();

assert_eq!(iterator.next(), Some(&1));
assert_eq!(iterator.next(), Some(&2));
assert_eq!(iterator.next(), Some(&4));
assert_eq!(iterator.next(), None);

pub fn insert(&mut self, index: usize, element: T) -> Result<(), T>

Insert the element into the list at index. Returns the element to be inserted if the index is out of bounds.

let mut list = btreelist![1, 2, 3];
list.insert(1, 4);
assert_eq!(list, btreelist![1, 4, 2, 3]);
list.insert(4, 5);
assert_eq!(list, btreelist![1, 4, 2, 3, 5]);

pub fn push(&mut self, element: T)

Push the element onto the back of the list.

let mut list = btreelist![1, 2];
list.push(3);
assert_eq!(list, btreelist![1, 2, 3]);

pub fn push_back(&mut self, element: T)

Push the element onto the back of the list.

let mut list = btreelist![1, 2];
list.push_back(3);
assert_eq!(list, btreelist![1, 2, 3]);

pub fn push_front(&mut self, element: T)

Push the element onto the front of the list.

let mut list = btreelist![2, 3];
list.push_front(1);
assert_eq!(list, btreelist![1, 2, 3]);

pub fn pop(&mut self) -> Option<T>

Remove and return the last element from the list, if there is one.

let mut list = btreelist![1, 2, 3];
assert_eq!(list.pop(), Some(3));
assert_eq!(list, btreelist![1, 2]);

pub fn pop_back(&mut self) -> Option<T>

Remove and return the last element from the list, if there is one.

let mut list = btreelist![1, 2, 3];
assert_eq!(list.pop_back(), Some(3));
assert_eq!(list, btreelist![1, 2]);

pub fn pop_front(&mut self) -> Option<T>

Remove and return the first element from the list, if there is one.

let mut list = btreelist![1, 2, 3];
assert_eq!(list.pop_front(), Some(1));
assert_eq!(list, btreelist![2, 3]);

pub fn remove(&mut self, index: usize) -> Option<T>

Removes the element at index from the list if it exists.

let mut list = btreelist![1, 2, 3];
assert_eq!(list.remove(1), Some(2));
assert_eq!(list, btreelist![1, 3]);

pub fn set(&mut self, index: usize, element: T) -> Result<T, T>

Update the element at index in the list, returning the old value on success, or the given value when the index is out of bounds.

let mut list = btreelist![1, 2, 3];
list.set(1, 4);
assert_eq!(list, btreelist![1, 4, 3]);

pub fn swap(&mut self, a: usize, b: usize) -> bool

Returns whether the swap was successful.

pub fn get(&self, index: usize) -> Option<&T>

Get the element at index in the list.

let list = btreelist![10, 40, 30];
assert_eq!(list.get(1), Some(&40));
assert_eq!(list.get(3), None);

pub fn get_mut(&mut self, index: usize) -> Option<&mut T>

Get the and element at index in the list.

let list = &mut btreelist![0, 1, 2];
if let Some(elem) = list.get_mut(1) {
    *elem = 42;
}
assert_eq!(*list, btreelist![0, 42, 2]);

pub fn first(&self) -> Option<&T>

Get the first element in the list if it exists.

let list = btreelist![10, 40, 30];
assert_eq!(list.first(), Some(&10));

pub fn first_mut(&mut self) -> Option<&mut T>

Get the first element in the list if it exists.

let list = &mut btreelist![0, 1, 2];
if let Some(elem) = list.first_mut() {
    *elem = 42;
}
assert_eq!(*list, btreelist![42, 1, 2]);

pub fn last(&self) -> Option<&T>

Get the last element in the list if it exists.

let list = btreelist![10, 40, 30];
assert_eq!(list.last(), Some(&30));

pub fn last_mut(&mut self) -> Option<&mut T>

Get the last element in the list if it exists.

let list = &mut btreelist![0, 1, 2];
if let Some(elem) = list.last_mut() {
    *elem = 42;
}
assert_eq!(*list, btreelist![0, 1, 42]);